requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==24.3
django-storages==1.14.4
django-unfold==0.47.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
grpcio==1.67.1
idna==3.10
inflection==0.5.1
milvus-lite==2.4.11
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pymilvus==2.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
